Organize files to be modular
