+++ src/App/App.css
@@ -1,0 +1,23 @@
+.App {
+  display: flex;
+  flex-direction: column;
+  background-color: #f6f6f7;
+  width: 100%;
+  height: 100%;
+  font-family:'Inter', sans-serif;
+}
+
+.header {
+  font-size: 24px;
+  color: #212223;
+  text-align: left;
+  padding: 0 0 0 5%
+}
+
+.subheader {
+  font-weight: 300;
+  font-size: 16px;
+  color:#6b7177;
+  padding: 0 0 0 5%;
+  margin: 0;
+}
