--- src/Loader.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import './Loader.css';
-
-const Loader = () => {
-  return (
-    <div className='loading-container'>
-    <h2 className='loading-header'>Please hold...</h2>
-    <p>Getting your images of Mars</p>
-    <img
-  )
-}
