+++ src/Loader/Loader.css
@@ -1,19 +1,25 @@
-.App {
+.loading-container {
   display: flex;
   flex-direction: column;
-  background-color: #f6f6f7;
-  width: 100%;
+  border-radius: 10px;
+  margin-top: 5%;
+  height: 100%;
+  width: 40%;
+  padding: 5%;
+  margin-left: 30%;
+  background-color: white;
+  box-shadow: 0 1px 5px rgb(0 0 0 / 20%);
   font-family:'Inter', sans-serif;
 }
 
-.header {
+.loading-header {
   font-size: 24px;
   color: #212223;
   text-align: left;
   padding: 0 0 0 5%
 }
 
-.subheader {
+.message-text {
   font-weight: 300;
   font-size: 16px;
   color:#7f8286;
